requirements/local-linux.txt
asgiref==3.4.1
autopep8==1.6.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.10
django-model-utils==4.2.0
djangorestframework==3.13.1
firebase-admin==5.2.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.5.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
msgpack==1.0.3
packaging==21.3
Pillow==8.4.0
proto-plus==1.19.9
protobuf==3.19.3
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.8
